refactor: split cli/main.py, core/verification.py, federation/groups.py (partial #180)
- cli/main.py (69KB → 12KB): extracted commands to cli/commands/ subpackage
- core/verification.py → core/verification/ package (constants, enums, models, reputation, service, validators)
- federation/groups.py → federation/groups/ package (admin, membership, mls)

All backward-compatible re-exports maintained.
